Fixed wrong `io.EOF` handling in sinker code
On Solana, using a block range where the end block number was not produced by the chain led to infinite retry of the stream.

We must assume that receiving an `io.EOF` means the stop block was reached. This is because on network that can skips block number, it's possible that we requested to stop on a block number that is no in the chain meaning we will receive `io.EOF` but the last seen block before it is not our block number, we must have confidence in the Substreams provider to respect the protocol